esri GRID raster support, wildcards support to include/exclude raster files.
Enhancements>
1. New templates (Grid_to_MRF_LERC, Grid_to_COG_LZW)

2. Wildcards support

OptimizeRasters now supports case sensitive UNIX style file-path search pattern support for the <RasterFormatFilter> and the <ExcludeFilter> in the configuration file. The raster search patterns only affect the raster processing and not for generating the orjob file that will continue to list all the raters with any metadata files found at the input source path. The rasters that do not fall into the search pattern criteria will be copied to the destination path without any processing done unless the extension or the appropriate search pattern exists to exclude any unwanted rasters in <ExcludeFilter>.
For e.g. to include only the raster World.TIF in the raster processing but to exclude all other TIF rasters found at the input source, update the relevant OptimizeRasters template file to update with <RasterFormatFilter>World.TIF</RasterFormatFilter> and <ExcludeFilter>*.TIF<ExcludeFilter>.

Please note the following usage cases to narrow down on the search results required before any raster processing is carried out,

Recognized search pattern notations,
*	Match any group of characters
?	Match any single character or match nothing if no character is found at the location in the filename.
[seq]	Matches any character in the square brackets
[!seq]	Matches any character not in the square brackets

Usage examples,

*.tif

Move the search to the end of the file-path to match for the ‘.tif’ extensions. Please note, this will not process extensions such as TIF, Tif, tiff, etc..

*Input/raster?/*.[Tt]if

The subfolder path can end with input/rasters or input/rasterz and the extension must end with either ‘tif’ or ‘Tif’

*/input[!0123]/*.jpg

The last subfolder in the file-path must not end with /input0, /input1, /input2 and /input3 and should look for the extension.jpg. The input4/raster.jpg file-path for e.g. is returned as a valid search result to process.

#82 #65