S3 Files >2GB Cause OverflowError Due to SSL
In python versions earlier than 3.10 the standard OpenSSL library version used only supported 32-bit reads. The result of this is that when trying to use OptimizeRasters on an S3 file that is > 2GB it fails with a non-intuitive error "signed integer greater than maximum value". The proposed code limits the amount of data that will be read in a single shot from S3 to avoid this issue. I have tested and verified this code against an image greater than 2GB.